Add a bunch of changes
I've tried to be careful to ensure that all the changes im pushing in
this commit hopefully wont break other envs when installed out of the
blue. Theres some env variables that I left in because I don't think
they matter. The stuff at the end of the zshrc might be trouble but I
plan on fixing that in a subsequent commit from my chromebook when I try
to set up programatic installation of ripgrep and fzf so and maybe tweek
the config for those so they ACTUALLY work in a useful way. My
plugins for vim are probably a little wacky. I was messing with them a
while back. Lets see how this works out!